refactor: centralize sync helpers
